in-queue multi threading
Queue constructor now accepts an argument to define in-queue threads. When >1, queued jobs don't wait for the previous one to finish but already start on a new thread if a free one is available.
The debugge now uses chalk to color the code as we also use it in the examples.
Also added an example to show the results.