log name
Replacing the log phrase 'klyft' with a variable that can be set manually when creating a new worker. When using false instead of a string, 'klyft' will be displayed.
To keep the examples working, the additional argument is set.
Attention: If you are updating to this version, you will need to update the worker options as your program might break otherwise!